+++ src/domain/git/queries/__tests__/detectConflict.test.ts
@@ -1,12 +1,13 @@
-import diff3Merge from "diff3"
 import fs from "fs"
 import * as git from "isomorphic-git"
+import path from "path"
+import pify from "pify"
 import assert from "power-assert"
 // tslint:disable-next-line
 import * as helpers from "../../__testHelpers__/helpers"
-import { getRefOids } from "../getRefOids"
+import { detectConflict } from "../detectConflict"
 
-test("hasConflict", async () => {
+test("no conflict", async () => {
   const dir = await helpers.createTempGitProject()
   await helpers.batchUpdateFiles(dir, [["x", "0"]])
 
@@ -16,36 +17,111 @@ test("hasConflict", async () => {
 
   await git.branch({ fs, dir, ref: "b" })
   await git.checkout({ fs, dir, ref: "b" })
-  await helpers.batchUpdateFiles(dir, [["x", "a\nb"]])
+  await helpers.batchUpdateFiles(dir, [["x", "b"]])
 
   await git.checkout({ fs, dir, ref: "a" })
   await git.branch({ fs, dir, ref: "c" })
   await git.checkout({ fs, dir, ref: "c" })
-  await helpers.batchUpdateFiles(dir, [["x", "a\nc"]])
+  await helpers.batchUpdateFiles(dir, [["x", "b"]])
 
-  const [our] = await getRefOids(dir, "a")
-  const [they1] = await getRefOids(dir, "b")
-  const [they2] = await getRefOids(dir, "c")
+  const ret = await detectConflict(dir, "a", "b", "c")
+  assert(ret.conflicts.length === 0)
+})
 
-  const { object: ours } = await git.readObject({ fs, dir, oid: our.oid })
-  const { object: t1 } = await git.readObject({
+test("with conflict", async () => {
+  const dir = await helpers.createTempGitProject()
+  await helpers.batchUpdateFiles(dir, [["x", "0"]])
+
+  await git.branch({ fs, dir, ref: "a" })
+  await git.checkout({ fs, dir, ref: "a" })
+  await helpers.batchUpdateFiles(dir, [["x", "a"]])
+
+  await git.branch({ fs, dir, ref: "b" })
+  await git.checkout({ fs, dir, ref: "b" })
+  await helpers.batchUpdateFiles(dir, [["x", "b"]])
+
+  await git.checkout({ fs, dir, ref: "a" })
+  await git.branch({ fs, dir, ref: "c" })
+  await git.checkout({ fs, dir, ref: "c" })
+  await helpers.batchUpdateFiles(dir, [["x", "c"]])
+
+  const ret = await detectConflict(dir, "a", "b", "c")
+  assert(ret.conflicts.length === 1)
+})
+
+test("no conflict with added files", async () => {
+  const dir = await helpers.createTempGitProject()
+  await helpers.batchUpdateFiles(dir, [["x", "0"]])
+
+  await git.branch({ fs, dir, ref: "a" })
+  await git.checkout({ fs, dir, ref: "a" })
+  await helpers.batchUpdateFiles(dir, [["x", "a"]])
+
+  await git.branch({ fs, dir, ref: "b" })
+  await git.checkout({ fs, dir, ref: "b" })
+  await helpers.batchUpdateFiles(dir, [["y", "1"]])
+
+  await git.checkout({ fs, dir, ref: "a" })
+  await git.branch({ fs, dir, ref: "c" })
+  await git.checkout({ fs, dir, ref: "c" })
+  await helpers.batchUpdateFiles(dir, [["y", "1"], ["z", "2"]])
+
+  const ret = await detectConflict(dir, "a", "b", "c")
+  assert(ret.conflicts.length === 0)
+})
+
+test("no conflict with removed file", async () => {
+  const dir = await helpers.createTempGitProject()
+  await helpers.batchUpdateFiles(dir, [["x", "0"]])
+
+  await git.branch({ fs, dir, ref: "a" })
+  await git.checkout({ fs, dir, ref: "a" })
+  await helpers.batchUpdateFiles(dir, [["x", "a"]])
+
+  await git.branch({ fs, dir, ref: "b" })
+  await git.checkout({ fs, dir, ref: "b" })
+  await pify(fs.unlink)(path.join(dir, "x"))
+  await git.remove({ fs, dir, filepath: "x" })
+  await git.commit({
     fs,
     dir,
-    oid: they1.oid
+    message: "Update",
+    author: { name: "n", email: "e" }
   })
 
-  const { object: t2 } = await git.readObject({
+  await git.checkout({ fs, dir, ref: "a" })
+  await git.branch({ fs, dir, ref: "c" })
+  await git.checkout({ fs, dir, ref: "c" })
+  await helpers.batchUpdateFiles(dir, [["y", "1"], ["z", "2"]])
+
+  const ret = await detectConflict(dir, "a", "b", "c")
+  assert(ret.conflicts.length === 0)
+})
+
+test("conflict with removed file", async () => {
+  const dir = await helpers.createTempGitProject()
+  await helpers.batchUpdateFiles(dir, [["x", "0"]])
+
+  await git.branch({ fs, dir, ref: "a" })
+  await git.checkout({ fs, dir, ref: "a" })
+  await helpers.batchUpdateFiles(dir, [["x", "a"]])
+
+  await git.branch({ fs, dir, ref: "b" })
+  await git.checkout({ fs, dir, ref: "b" })
+  await pify(fs.unlink)(path.join(dir, "x"))
+  await git.remove({ fs, dir, filepath: "x" })
+  await git.commit({
     fs,
     dir,
-    oid: they2.oid
+    message: "Update",
+    author: { name: "n", email: "e" }
   })
 
-  const diff3 = diff3Merge(
-    ours.toString().split("\n"),
-    t1.toString().split("\n"),
-    t2.toString().split("\n")
-  )
+  await git.checkout({ fs, dir, ref: "a" })
+  await git.branch({ fs, dir, ref: "c" })
+  await git.checkout({ fs, dir, ref: "c" })
+  await helpers.batchUpdateFiles(dir, [["x", "c"]])
 
-  const hasConflict = diff3.some((i: any) => i.conflict)
-  assert(hasConflict)
+  const ret = await detectConflict(dir, "a", "b", "c")
+  assert(ret.conflicts.length === 1)
 })